Instead of manually handling demangling, for cross-platform compatibility `boost`'s `typeindex` library have been used.
Templated classes such as `BaseAnimComp` weren't displaying it's single name. It's been fixed.
