+++ CMakeLists.txt
@@ -42,10 +42,12 @@ add_executable(ETG main.cpp
         src/Utils/Math.h
         src/Managers/SpriteBatch.cpp
         src/Core/Components/BaseAnimComp.h
-        src/Engine/UI/EngineUI.cpp
-        src/Engine/UI/EngineUI.h
+        src/Engine/Engine.cpp
+        src/Engine/Engine.h
         src/Core/GameObject.cpp
         src/Managers/InputManager.cpp
+        src/Engine/UI/EngineUI.cpp
+        src/Engine/UI/EngineUI.h
 )
 
 # Easiest and fastest way is VCPKG 
@@ -53,6 +55,7 @@ find_package(SFML CONFIG REQUIRED system window graphics network audio)
 find_package(benchmark CONFIG REQUIRED)
 find_package(Threads REQUIRED)
 find_package(ImGui-SFML CONFIG REQUIRED)
+find_package(boost_type_index CONFIG REQUIRED)
 
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang" AND "${CMAKE_CXX_COMPILER_FRONTEND_VARIANT}" STREQUAL "MSVC")
         set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /EHsc")
@@ -75,6 +78,7 @@ target_link_libraries(ETG PRIVATE
         sfml-window
         benchmark::benchmark
         ImGui-SFML::ImGui-SFML
+        Boost::type_index
 )
 
 install(TARGETS ETG
