+++ src/Core/Components/BaseAnimComp.h
@@ -6,15 +6,15 @@ namespace ETG
 {
     enum class HeroStateEnum;
 
-    template<typename StateEnum>
-    class BaseAnimComp : public GameObject<BaseAnimComp<StateEnum>>
+    template<typename Derived, typename StateEnum>
+    class BaseAnimComp : public GameObject<Derived>
     {
     public:
-        BaseAnimComp();
+        BaseAnimComp() = default;
         virtual void SetAnimations();
         void Update(const StateEnum& stateEnum, const AnimationKey& animKey);
         virtual void Draw(const sf::Vector2f& position);
-        virtual void Draw(const sf::Vector2f position, const sf::Vector2f Origin, const sf::Vector2f Scale, const float Rotation, const float depth);
+        virtual void Draw(sf::Vector2f position, sf::Vector2f Origin, sf::Vector2f Scale, float Rotation, float depth);
         
         sf::IntRect CurrTexRect;
         sf::Vector2f RelativeOrigin{0.f, 0.f};
@@ -26,8 +26,8 @@ namespace ETG
         AnimationKey CurrentAnimStateKey;
     };
 
-    template <typename StateEnum>
-    void BaseAnimComp<StateEnum>::Update(const StateEnum& stateEnum, const AnimationKey& animKey)
+    template <typename DerivedName, typename StateEnum>
+    void BaseAnimComp<DerivedName, StateEnum>::Update(const StateEnum& stateEnum, const AnimationKey& animKey)
     {
         CurrentState = stateEnum;
         CurrentAnimStateKey = animKey;
@@ -43,31 +43,24 @@ namespace ETG
         RelativeOrigin = animManager.AnimationDict[AnimManagerDict[CurrentState].LastKey].Origin;
     }
 
-    template <typename StateEnum>
-    void BaseAnimComp<StateEnum>::Draw(const sf::Vector2f& position)
+    template <typename Derived, typename StateEnum>
+    void BaseAnimComp<Derived, StateEnum>::Draw(const sf::Vector2f& position)
     {
         if (!AnimManagerDict.contains(CurrentState)) throw std::runtime_error("AnimManagerDict doesn't contain given state");
 
         AnimManagerDict[CurrentState].Draw(CurrentTex, position, sf::Color::White, this->Rotation, RelativeOrigin, this->Scale, this->Depth);
     }
 
-    template <typename StateEnum>
-    void BaseAnimComp<StateEnum>::Draw(const sf::Vector2f position, const sf::Vector2f Origin, const sf::Vector2f Scale, const float Rotation, const float depth)
+    template <typename Derived, typename StateEnum>
+    void BaseAnimComp<Derived, StateEnum>::Draw(const sf::Vector2f position, const sf::Vector2f Origin, const sf::Vector2f Scale, const float Rotation, const float depth)
     {
         if (!AnimManagerDict.contains(CurrentState)) throw std::runtime_error("AnimManagerDict doesn't contain given state");
 
         AnimManagerDict[CurrentState].Draw(CurrentTex, position, sf::Color::White, Rotation, Origin, Scale, depth);
     }
 
-    template <typename StateEnum>
-    BaseAnimComp<StateEnum>::BaseAnimComp()
+    template <typename Derived, typename StateEnum>
+    void BaseAnimComp<Derived, StateEnum>::SetAnimations()
     {
-        //Empty
-    }
-
-    template <typename StateEnum>
-    void BaseAnimComp<StateEnum>::SetAnimations()
-    {
-        //empty
     }
 }
