+++ src/Engine/UI/EngineUI.cpp
@@ -1,131 +1,62 @@
+//
+// Created by selviniah on 23/02/25.
+//
+
 #include "EngineUI.h"
-#include "../../Managers/GameManager.h"
-#include <imgui-SFML.h>
+#include "../../Core/GameObject.h"
 #include <imgui.h>
-#include <iostream>
 
-bool EngineUI::CurrentGameFocus = false;
-bool EngineUI::PreviousGameFocus = false;
-
-using namespace ETG;
-
-void EngineUI::Initialize()
+void ImGuiSetRelativeOrientation(GameObjectBase* obj)
 {
-    //Initialize Imgui-SFML after creating the window
-    if (!ImGui::SFML::Init(*ETG::Window)) throw std::runtime_error("Cannot initialize ImGUI with the given Window");
-
-    ETG::GameState::GetInstance().SetEngineUISize(windowSize);
-    windowSize = {300, (float)(ETG::Window->getSize().y)};
-    std::cout << std::unitbuf;
-
-    LoadFont();
-}
-
-void EngineUI::Update() const
-{
-    //Update ImGUI-SFML with the frame delta time
-    ImGui::SFML::Update(*ETG::Window, ETG::ElapsedTimeClock);
-
-    // Begin a new ImGui window docked to the right
-    ImGui::SetNextWindowPos(ImVec2((float)ETG::Window->getSize().x - windowSize.x, 0));
-    ImGui::SetNextWindowSize(ImVec2(windowSize));
-
-    ImGui::Begin("Details Pane", nullptr);
-    UpdateDetailsPanel(ETG::GameState::GetInstance().GetSceneObj());
-
-    PreviousGameFocus = CurrentGameFocus;
-    CurrentGameFocus = !(ImGui::IsWindowHovered() || ImGui::IsWindowFocused());
-
-    //The end
-    ImGui::End();
-}
-
-bool EngineUI::IsGameWindowFocused()
-{
-    // Check if the game has focus; if not, ignore input.
-    if (!CurrentGameFocus) return false;
-
-    // Determine if focus was just gained (Current is true, Previous was false)
-    if (CurrentGameFocus && !EngineUI::PreviousGameFocus)
-        InputManager::LeftClickRequired = true;
-
-    // Process events to check for mouse release
-    if (InputManager::LeftClickRequired && GameManager::GameEvent.type == sf::Event::MouseButtonReleased && ETG::GameManager::GameEvent.mouseButton.button == sf::Mouse::Left)
-    {
-        InputManager::LeftClickRequired = false;
-    }
     
-    // If LeftClickRequired is still true, wait for release before processing input
-    if (InputManager::LeftClickRequired)
+    //NOTE: Position
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Position");
+    ImGui::SameLine();
+    if (ImGui::InputFloat2("##PosX", &obj->RelativePos.x, "%.3f"))
     {
-        return false;
-    }
-    return true;
-}
-
-void EngineUI::Draw()
-{
-    //Render
-    ImGui::SFML::Render(*ETG::Globals::Window);
-}
-
-void EngineUI::UpdateDetailsPanel(const std::vector<GameObjectBase*>& SceneObjects)
-{
-    static int SelectedIdx = 0;
-
-    //NOTE: Open the Game Objects pane by default
-    ImGui::SetNextItemOpen(true,ImGuiCond_Once);
-    
-    //List all objects
-    if (ImGui::CollapsingHeader("Game Objects", ImGuiTreeNodeFlags_None))
-    {
-        for (int i = 0; i < SceneObjects.size(); ++i)
-        {
-            ImGui::PushID(i);
-
-            //The boolean only for highlighting selection
-            const bool isSelected = (SelectedIdx == i);
-            
-            if (ImGui::Selectable(SceneObjects[i]->GetObjectName().c_str(), isSelected))
-                SelectedIdx = i;
-
-            ImGui::PopID();
-        }
+        //Revert what I've done last time and then apply RelativePos again
+        const sf::Vector2f delta = {obj->RelativePos.x - obj->PrevRelativePos.x, obj->RelativePos.y - obj->PrevRelativePos.y};
+        obj->DrawProps.Position += delta;
+        obj->PrevRelativePos = obj->RelativePos;
     }
 
-    //NOTE: Open the details pane by default
-    ImGui::SetNextItemOpen(true,ImGuiCond_Once);
-
-    //Show details of selected objects
-    if (ImGui::CollapsingHeader("Details", ImGuiTreeNodeFlags_None))
+    //NOTE: Scale
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Scale");
+    ImGui::SameLine();
+    if (ImGui::InputFloat2("##Scale", &obj->RelativeScale.x, "%.3f"))
     {
-        //Open the pane by default once
-        ImGui::SetNextItemOpen(true,ImGuiCond_Once);
-        
-        if (SelectedIdx >= 0 && SelectedIdx < SceneObjects.size())
-        {
-            if(ImGui::TreeNode("Absolute Orientation"))
-            {
-                SceneObjects[SelectedIdx]->ImGuiSetAbsoluteOrientation();
-                ImGui::TreePop();   
-            }
-
-            if(ImGui::TreeNode("Relative Orientation"))
-            {
-                SceneObjects[SelectedIdx]->ImGuiSetRelativeOrientation();
-                ImGui::TreePop();   
-            }
-        }
+        const sf::Vector2f delta{obj->RelativeScale.x - obj->PrevRelativeScale.x, obj->RelativeScale.y - obj->PrevRelativeScale.y};
+        obj->DrawProps.Scale += delta;
+        obj->PrevRelativeScale = obj->RelativeScale;
     }
+
+    //NOTE: Rotation
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Rotation");
+    ImGui::SameLine();
+    ImGui::SliderFloat("##Rot", &obj->RelativeRotation, 0.0f, 360.0f);
 }
 
-void EngineUI::LoadFont()
+// When the UI changes, recalculate the final position from BasePosition.
+void ImGuiSetAbsoluteOrientation(GameObjectBase* obj)
 {
-    const ImGuiIO io = ImGui::GetIO();
-    io.Fonts->Clear();
+    //NOTE: Position
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Position");
+    ImGui::SameLine();
+    if (ImGui::InputFloat2("##PosX", &obj->Position.x, "%.3f")) {}
 
-    const std::filesystem::path FontPath = std::filesystem::path(RESOURCE_PATH) / "Fonts" / "SegoeUI.ttf";
-    SegoeFont = io.Fonts->AddFontFromFileTTF(FontPath.generic_string().data(), 18.f);
-    if (SegoeFont == nullptr) throw std::runtime_error("Failed to load font from " + FontPath.generic_string());
-    if (!ImGui::SFML::UpdateFontTexture()) throw std::runtime_error("Font cannot be load.");
+    //NOTE: Scale
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Scale");
+    ImGui::SameLine();
+    if (ImGui::InputFloat2("##Scale", &obj->Scale.x, "%.3f")) {}
+
+    //NOTE: Rotation
+    ImGui::AlignTextToFramePadding();
+    ImGui::Text("Rotation");
+    ImGui::SameLine();
+    ImGui::SliderFloat("##Rot", &obj->Rotation, 0.0f, 360.0f);
 }
